Stop pulling X11 stack into hypervisor
Those packages are not of general use, and there is no place for them on
a headless hypervisor. If someone needs them to get a local GUI, fine,
they can install the packages as usual, but that should not be a
requirement for Infrared deployments.

Change-Id: Ib56f4f504770a4266aaef433ae9fa8d799ce7aaf
(cherry picked from commit a58eef4d7c201fc25c055a50977c95eb3a859661)