Connectivity to the hybrid mode
Allows to connect to the overcloud nodes thorough the hypervisor (1 hop).
In order to do that the hypervisor should be connected to
the provisioner network. This can be done by:
 1) assigning static ip for the br-ctlplane bridge
 2) using dhcp for the br-ctlplane bridge
All that can be done by editing network topology config file and adding
'ip_address' or 'bootproto' options to the network description.

Change-Id: Iba5456a1391aab7692a606bad0d32b9f0f0099d2