Add virsh node parameter to specify if image should be deployed
This change adds the option in the virsh nodes definition to specify
whether the OS gets deployed or not. The current condition is to
not deploy the image if the node is part of the overcloud_nodes group.
In case of split stack deployments we need the overcloud nodes to get
the image deployed as they don't get provisioned by Ironic during
the overcloud deployment phase.

Change-Id: Ie2283bf4b3d0a9b556e8e42b65bd2f9c744a6162