Execute dhclient to recover access to the interface
When we are connected to the interface (for example, because is
used for ssh access), this command causes to loose access to it.
Execute a dhclient to regain inmediate access, and continue using
it without disruption.

Also we need to bring down the bridge before bringing down
interface, otherwise this is throwing an error for this
specific use case of reusing the interface.

Change-Id: I6740a8f0be4d879d5f23af10c943666120ca5a8e
Authored-By: Yolanda Robla <yroblamo@redhat.com>
Co-Authored-By: Zenghui Shi <zshi@redhat.com>