Skip authorized_key on cleanup
Virsh cleanup can fail on autorized_key task from provision_user.yml
when id_rsa.pub key is not present in the workspace. This can happen
for example when we use the same worksapce but run different provisioner's.
For example run beacker provision and then run virsh cleanup.

The problem is that ansiblie resolves with_file loop every time and apply
'when' condition for every file. And so the authorized_key fails because
required file cannot be found.

During cleanup the pub key is not required so we should not fail in that case.
This patch removes with_file argument and uses lookup call instead.

Change-Id: Ibd2b4731a773c9a15c92ff4702c72eda25a256fa