Add 4_nets_3_bridges_hybrid network in virsh plugin
Add a support for the hybrid deployments in virsh plugin.
Each network is using a real interface of the hypervisor and serves a
dedicated role - Provision / API / DataPlane.

In case the name of the interface differ from the name present in the
hypervisor, the value of the network or any other value could be
overriden in the following way:
-e override.networks.net1.nic=eno2

Change-Id: I5e7aeec30bb0557d77b4f72c596c4f0cdc458806