virsh: Add support for creating virtual switches
The patch implements support for creation of a virtual switch including
wiring with nodes. So far only vqfx switch is implemented but it will be
easy to extend support for different vendors. There is a new specific
task that configures switch after it's installed using expect commands
as libguestfs does not support vqfx operating system.

Because of that, a new node variable supports_libguestfs was introduced.
If the variable is set to false, an image is not provisioned by
infrared. If os.type of a node is set to vqfx a configuration script for
vqfx is executed.

Because vqfx does not support virtio devices, all network interfaces
must be set to e1000 and disks must be ide type. This also required a
change so disk type is not configurable per disk and not only per whole
environment.

In the added network topology, there is always a single linux bridge set
to hub mode between each "wire" from overcloud node to the vswitch.

Change-Id: If0bb92e528ccfd4814b23236e8fb7589b9f28fc6